requirements.txt
langchain-core==0.1.13
streamlit==1.30.0
langchain==0.1.1
langchain-openai==0.0.3
langchain-community==0.0.13
faiss-cpu==1.7.4
transformers==4.36.2
torch==2.1.2
accelerate==0.26.1
pinecone-client==3.0.3
langchain-pinecone==0.0.2
sentence-transformers==2.3.1
